refactor: move more grid functions to grid.c. Clean up some variables
